♻️  refactor: external 모듈로 분리, FB동작과 걸음 수 동작에 대한 서비스 분리 #121
