+++ motivoo-api/src/test/java/controller/BaseControllerTest.java
@@ -27,8 +27,8 @@ import sopt.org.motivoo.domain.auth.config.jwt.JwtTokenProvider;
 import sopt.org.motivoo.domain.auth.repository.TokenRedisRepository;
 import sopt.org.motivoo.external.client.auth.apple.service.AppleClaimsValidator;
 import sopt.org.motivoo.external.client.auth.apple.service.AppleLoginService;
-import sopt.org.motivoo.domain.external.firebase.FirebaseService;
-import sopt.org.motivoo.domain.external.firebase.config.FirebaseConfig;
+import sopt.org.motivoo.external.firebase.FirebaseService;
+import sopt.org.motivoo.external.firebase.config.FirebaseConfig;
 import sopt.org.motivoo.external.s3.S3Service;
 import sopt.org.motivoo.external.s3.config.AWSConfig;
 import sopt.org.motivoo.domain.external.slack.SlackService;
